requirements.txt
Django>=4.2
djangorestframework>=3.14
mongoengine==0.29.1
PyJWT>=2.8
python-dotenv>=1.0
django-cors-headers>=4.0
Pillow>=10.0
werkzeug>=2.0
gunicorn>=20.0
dnspython>=2.3.0
